chore: isolate CodSpeed comparison benchmarks
