+++ .github/workflows/test.yml
@@ -150,7 +150,15 @@ jobs:
       - name: "Build benchmarks"
         run: |
           cargo codspeed build -m simulation -m memory \
-            --bench compare \
+            --bench compare_amortized_input \
+            --bench compare_amortized_interned_input \
+            --bench compare_amortized_supertype_input \
+            --bench compare_cached_input \
+            --bench compare_mutating_inputs \
+            --bench compare_new_input \
+            --bench compare_new_interned_input \
+            --bench compare_new_supertype_input \
+            --bench compare_read_field \
             --bench incremental \
             --bench accumulator \
             --bench dataflow
