+++ benches/compare/support/interned.rs
@@ -1,0 +1,11 @@
+#[salsa::interned]
+pub struct InternedInput<'db> {
+    #[returns(ref)]
+    pub text: String,
+}
+
+#[salsa::tracked(returns(copy))]
+#[inline(never)]
+pub fn interned_length<'db>(db: &'db dyn salsa::Database, input: InternedInput<'db>) -> usize {
+    input.text(db).len()
+}
